+++ rook_receiver/app/src/main/java/au/com/ldcs/flexdisplay/rook/BootReceiver.java
@@ -1,0 +1,14 @@
+package au.com.ldcs.flexdisplay.rook;
+
+import android.content.BroadcastReceiver;
+import android.content.Context;
+import android.content.Intent;
+
+public final class BootReceiver extends BroadcastReceiver {
+    @Override
+    public void onReceive(Context context, Intent intent) {
+        Intent launch = new Intent(context, MainActivity.class)
+                .addFlags(Intent.FLAG_ACTIVITY_NEW_TASK | Intent.FLAG_ACTIVITY_CLEAR_TOP);
+        context.startActivity(launch);
+    }
+}